Endianness reversing methods
Added Reverse methods to Binary16, Binary32 and Binary64 types which return a new instance with reversed byte order.
Added constructors for the aforementioned types which take a byte pointer.
Added static BinaryUtility class with methods which reverse the endianness of arbitrary data, defined as a byte pointer, byte array, or numerous byte-reference arguments.